+++ bridges/otellogr/logsink.go
@@ -176,6 +176,7 @@ func NewLogSink(name string, options ...Option) *LogSink {
 		logger:        c.provider.Logger(name, opts...),
 		levelSeverity: c.levelSeverity,
 		opts:          opts,
+		ctx:           context.Background(),
 	}
 }
 
@@ -240,7 +241,7 @@ func (l *LogSink) Info(level int, msg string, keysAndValues ...any) {
 
 // Init receives optional information about the logr library this
 // implementation does not use it.
-func (l *LogSink) Init(info logr.RuntimeInfo) {
+func (l *LogSink) Init(logr.RuntimeInfo) {
 	// We don't need to do anything here.
 	// CallDepth is used to calculate the caller's PC.
 	// PC is dropped as part of the conversion to the OpenTelemetry log.Record.